Migrate numeric geometry types from f32 to f64
Switch core numeric types from f32 to f64 across the codebase to improve numerical precision. Changes touch geometry primitives (Point, Rect, Circle, Arc, Edge, BoundarySegment, SPolygon), geometry utilities (convex hull, DTransformation, fail-fast surrogates/poles/piers), traits (DistanceTo/SeparationDistance signatures), collision/quadtree routines, entities (area/density functions), layout snapshot, and various IO/problem modules. This is an API-breaking numeric precision migration — method signatures, structs and trait return types changed to f64 and callers may need updates.

+++ src/collision_detection/cd_engine.rs
@@ -471,7 +471,7 @@ mod tests {
         SPolygon::new(points).unwrap()
     }
 
-    fn rect_query(x_min: f32, y_min: f32, x_max: f32, y_max: f32) -> SPolygon {
+    fn rect_query(x_min: f64, y_min: f64, x_max: f64, y_max: f64) -> SPolygon {
         SPolygon::from(Rect::try_new(x_min, y_min, x_max, y_max).unwrap())
     }
